requirements.txt
Django==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
gunicorn==19.9.0
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
torch==0.4.1
torchvision==0.2.1
